update: Add Try-Catch and -TaskCredentials
the -TaskCredentials parameter is required to be passed with the SCOM Start-SCOMTask cmdlet.

As per the 2022 documentation:  This is a mandatory parameter. If you specify a null value, Operations Manager uses the default TaskCredentials of the account for the current user.

2022:
https://learn.microsoft.com/en-us/powershell/module/operationsmanager/start-scomtask?view=systemcenter-ps-2022

In prior versions, if you do not specify this parameter, the default is the account for the current user.

2019:
https://learn.microsoft.com/en-us/powershell/module/operationsmanager/start-scomtask?view=systemcenter-ps-2019